code/BE/pyproject.toml
[tool.poetry]
name = "final_server"
version = "0.1.0"
description = "final project server with fast-api"
authors = ["woowonjin <dndnjswls613@naver.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = ">=3.7.1, <3.10"
fastapi = "^0.70.0"
uvicorn = "^0.15.0"
tqdm = "^4.62.3"
pandas = "^1.3.5"
datasets = "^1.16.1"
torch = "^1.10.1"
transformers = "^4.14.1"
pydantic = "^1.8.2"
scikit-learn = "^1.0.1"
python-dotenv = "^0.19.2"
google-cloud-storage = "^1.43.0"
google-api-python-client = "^2.33.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
black = "^21.12b0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
